Revert "Webpack task runner assignment"
