--- .gitignore
@@ -1,8 +1,0 @@
-/node_modules
-/dist
-.DS_Store
-*.log
-.idea
-.vscode
-.cache
-.eslintcache
