+++ assets/js/components/author-card.js
