+++ authors.html
@@ -7,11 +7,11 @@
     <meta name="author" content="Bryan Fernandez">
     <meta name="description" content="This is a blog about technology, where everybody has access to post  any content related to.">
     <link rel="stylesheet" href="https://pro.fontawesome.com/releases/v5.10.0/css/all.css" integrity="sha384-AYmEC3Yw5cVb3ZcuHtOA93w35dYTsvhLPVnYs9eStHfGJvOvKxVfELGroGkvsg+p" crossorigin="anonymous"/>
+    <link rel="stylesheet" href="./assets/css/styles.css">
     <title>TechnoCrazy</title>
 </head>
 <body>
-    <div class="theme theme--default">
-         <!-- HEADER SECTION -->
+    <!-- HEADER SECTION -->
     <header class="header header--default">
         <div class="container">
             <nav class="nav">
@@ -20,7 +20,7 @@
                     <a href="#" class="nav__brand">Technocrazy</a>
                     <ul class="nav__list">
                         <li class="nav__item">
-                            <a href="../index.html" class="nav__link">Home</a>
+                            <a href="./index.html" class="nav__link">Home</a>
                         </li>
                         <li class="nav__item">
                             <a href="#" class="nav__link">Reviews</a>
@@ -44,10 +44,10 @@
                     </div>
                     <ul class="nav__list">
                         <li class="nav__item">
-                            <a href="./signIn.html" class="nav__link">Sign In</a>
+                            <a href="#" class="nav__link nav__link--active">Sign In</a>
                         </li>
                         <li class="nav__item">
-                            <a href="#" class="btn btn--primary">Sign Up</a>
+                            <a href="#" class="btn btn--outline-primary">Sign Up</a>
                         </li>
                     </ul>      
                 </div>
@@ -87,42 +87,44 @@
     <!-- MAIN CONTENT -->
 
     <main>
-        <section class="container container--centered">
-            <section class="auth-card">
-                <div class="auth-card__content">
-                    <header class="auth-card__content-header">
-                        <h1 class="auth-card__title">Sign Up</h1>
-                    </header>
-                    <form id="register_form" class="auth-form">
-                        <fieldset>
-                            <div class="auth-form__input-group">
-                                <label for="name" class="auth-form__input-label">Name</label>
-                                <input id="name" type="text" class="auth-form__input"  placeholder="juan Perez">
-                                <span class="auth-form__error-message"></span>
-                            </div>    
-                            <div class="auth-form__input-group">
-                                <label for="email" class="auth-form__input-label">Email</label>
-                                <input id="email" type="text" class="auth-form__input"  placeholder="juan.perez@example.com">
-                                <span class="auth-form__error-message"></span>
-                            </div>    
-                            <div class="auth-form__input-group">
-                                <label for="password" class="auth-form__input-label">Password</label>
-                                <input id="password" type="password" class="auth-form__input">
-                                <span class="auth-form__error-message"></span>
-                            </div>  
-                            <div class="auth-form__input-group">
-                                <label for="password" class="auth-form__input-label">Password Confirmation</label>
-                                <input id="password_validation" type="password" class="auth-form__input">
-                                <span class="auth-form__error-message"></span>
-                            </div>  
-                            <div class="auth-form__input-group">
-                                <button  type="submit" class="btn btn--primary btn--block">Register</button>
-                            </div>     
-                        </fieldset>
-                    </form>
-                </div>
-            </section>
-        </section>
+        <div class="container">
+            <div class="list-card">
+                <author-card 
+                    class="list-card__item"
+                    name="Bryan Flores"
+                    photo="https://images.pexels.com/photos/220453/pexels-photo-220453.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940"
+                    >
+                    <div slot="info">
+                        <p>Posts: 15</p>
+                        <p>Followers: 16</p>
+                        <p>Email: Bryan@exels.com</p>
+                    </div>
+                </author-card>
+                <author-card 
+                    class="list-card__item"
+                    name="Juan Perez"
+                    photo="https://wallpaperaccess.com/full/2213424.jpg"
+                    >
+                    <div slot="info">
+                        <p>Posts: 25</p>
+                        <p>Followers: 256</p>
+                        <p>Email: Juan@exels.com</p>
+                    </div>
+                </author-card>
+                <author-card 
+                    class="list-card__item"
+                    name="Xavier Rodriguez"
+                    photo="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRij6dtiHizH96qpCOe8WeXXP3yLyQJkPdGVg&usqp=CAU"
+                    >
+                    <div slot="info">
+                        <p>Posts: 95</p>
+                        <p>Followers: 1150</p>
+                        <p>Email: Xavier@gmail.com</p>
+                    </div>
+                </author-card>
+            </div>
+            
+        </div>
     </main>
 
     <!-- FOOTER SECTION -->
@@ -166,7 +168,7 @@
                         <h5 class="footer__title">Others</h5>
                         <ul>
                             <li><a href="#" class="footer__link">Reviews</a></li>
-                            <li><a href="./authors.html" class="footer__link">Authors</a></li>
+                            <li><a href="#" class="footer__link">Authors</a></li>
                             <li><a href="#" class="footer__link">Posts</a></li>
                             <li><a href="#" class="footer__link">News</a></li>
                         </ul>
@@ -192,11 +194,10 @@
                 </div>
             </div>
     </footer>
-    </div>
-   
     
    
-    <!-- <script type="module" src="./assets/js/main.js"></script>
-    <script type="module" src="./assets/js/signUp.js"></script> -->
+    <script type="module" src="./assets/js/main.js"></script>
+    <script src="./assets/js/components/author-card.js"></script>
+    
 </body>
 </html>
