--- package.json
@@ -1,47 +1,0 @@
-{
-  "name": "frontend_development",
-  "version": "1.0.0",
-  "description": "",
-  "private": true,
-  "scripts": {
-    "build": "rimraf dist && webpack",
-    "watch": "webpack --watch",
-    "serve": "node server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/bryanff/fronent_module.git"
-  },
-  "author": "Bryan Fernandez",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/bryanff/fronent_module/issues"
-  },
-  "homepage": "https://github.com/bryanff/fronent_module#readme",
-  "devDependencies": {
-    "@babel/core": "^7.17.4",
-    "@babel/preset-env": "^7.16.11",
-    "babel-loader": "^8.2.3",
-    "css-loader": "^6.6.0",
-    "css-minimizer-webpack-plugin": "^3.4.1",
-    "eslint": "^8.9.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.25.4",
-    "file-loader": "^6.2.0",
-    "html-loader": "^3.1.0",
-    "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.5.3",
-    "rimraf": "^3.0.2",
-    "sass": "^1.49.8",
-    "sass-loader": "^12.6.0",
-    "style-loader": "^3.3.1",
-    "webpack": "^5.69.1",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4"
-  },
-  "dependencies": {
-    "@babel/plugin-transform-runtime": "^7.17.0",
-    "@babel/runtime": "^7.17.2"
-  }
-}
