--- src/sass/_footer.scss
@@ -1,62 +1,0 @@
-.footer{
-    margin-top: auto;
-    background-color: $footer-color;
-    &__wrapper{
-        display: flex;
-        justify-content: space-evenly;  
-        flex-wrap: wrap;
-    }
-    &__content{
-        min-width: 250px;
-        margin-bottom: 30px;
-    }
-    &__bottom{
-        border-top: 1px solid #283044;
-        display: flex;
-        flex-wrap: wrap;
-        justify-content:space-around;
-        padding: 20px;
-        margin-top: 20px;
-    
-    }
-    &__link{
-        text-decoration: none;
-        color: $gray-color;
-        margin-right: 10px;
-        font-size: 1rem;
-        &--social{
-            font-size: 1.5rem;
-            &:hover{
-                transition: all 0.4s ease;
-                color: $primary-color-default;
-            }
-        }
-    }
-    &__title{
-        font-size: 1.2rem;
-        color: $white-color;
-        padding: 20px 0
-    }
-    &__copyright{
-        color: $text-color;
-    }
-}
-.subscribe-form{
-    &__email{
-        background-color: #272c30;
-        border-radius: 2px;
-        border: 1px solid #283044;
-        width: 200px;
-        height: 25px;
-        padding:10px;
-    }
-    &__button{
-        background-color: $primary-color-default;
-        border: 1px solid #283044;
-        color: $white-color;
-        padding: 3px 6px;
-        border-radius: 2px;
-        cursor: pointer;
-    }
-}
-
