--- src/sass/_main.scss
@@ -1,9 +1,0 @@
-@import "./variables";
-@import "./theme";
-@import "./base";
-@import "./navigation";
-@import "./components/card";
-@import "./components/button";
-@import "./components/stickyshare";
-@import "./components/tags";
-@import "./footer";
