--- src/sass/_variables.scss
@@ -1,39 +1,0 @@
-$content-color:"content-color";
-$sidebar-color:"sidebar-color";
-$body-color:"body-color";
-$nav-color:"nav-color";
-$side-nav: "side-nav";
-$search-bar:"search-bar";
-$text-color:"text-color";
-$primary-color:"primary-color";
-
-$content-color-default: #fff;
-$sidebar-color-default: #fff;
-$body-color-default: #f0f1f5;
-$nav-color-default: #e4e9f7;
-$side-nav-default: f4f5f9;
-$search-bar-default: #f2f2f2;
-$text-color-default: #506172;
-$primary-color-default: #ef3f48;
-
-$z-index-fixed: 100;
-
-$footer-color: #010f22;
-$gray-color: #a8a8a8;
-$red-color: #ff2d55;
-$blue-color: #007aff;
-$white-color: #f0f1f5;
-
-$content-color-dark:#34363c;
-$sidebar-color-dark:#34363c;
-$body-color-dark:#202124f2;
-$nav-color-dark:#242526;
-$side-nav-dark: #010718;
-$search-bar-dark:#f2f2f2;
-$text-color-dark: #bdc1c6;
-$primary-color-dark: #ef3f48;
-
-$main-font: "Playfair Display", "Segoe UI", "Helvetica Neue", "Arial";
-$small: 768px;
-$medium: 925px;
-$large: 100vw;
