--- src/sass/components/_stickyshare.scss
@@ -1,50 +1,0 @@
-
-.stickyshare {
-    text-align: center;
-    flex-direction: column;
-    align-items: center;
-    position: fixed;
-    top: 30%;
-    left: 25px;
-    z-index: 3;
-    @media screen and (max-width: $medium) {
-        display:none;
-    }
-    &__rotate {
-        margin-top: 0;
-        margin-bottom: 40px;
-        margin-left: 0;
-        margin-right: -17px;
-        align-items: center;
-        transform: rotate(-90deg);
-        transform-origin: 50% 100%;
-    }
-    &__line{
-        width:1px;
-        height:30px;
-        margin: 0px auto;
-        margin-bottom:20px;
-        @include themed() {
-            background-color: t($text-color);
-        }
-    }
-    &__social-link {
-        display: block;
-        font-size: 1rem;
-        @include themed() {
-            color: t($text-color);
-        }
-        margin: 10px auto;
-    }
-    &__text {
-        display: inline-block;
-        margin-bottom: -3px;
-        @include themed() {
-            color: t($text-color);
-        }
-        font-size: 1rem;
-        border-bottom: 1px solid #333;
-        font-family: Arial;
-        border-bottom:0;
-    }
-}
