--- src/sass/components/_tags.scss
@@ -1,20 +1,0 @@
-@mixin tagColor($background, $color) {
-    background-color: $background;
-    color: $color;
-}
-
-.tag{
-    padding: 4px 10px 4px 19px;
-    font-size: .75rem;
-    font-weight: bold;
-    text-transform: uppercase;
-    border-radius: 10px;
-    position: relative;
-    margin: 50px 0;
-    &--red{
-        @include tagColor($red-color, $white-color)
-    }
-    &--blue{
-        @include tagColor($blue-color, $white-color)
-    }
-}
